fix Order struct Order.accrual type

+++ models/models.go
@@ -13,11 +13,11 @@ type UserOrder struct {
 }
 
 type Order struct {
-	ID         int    `json:"-"`
-	Number     int    `json:"number"`
-	Status     string `json:"status"`
-	Accrual    int    `json:"accrual"`
-	UploadedAt string `json:"uploaded_at"`
+	ID         int     `json:"-"`
+	Number     int     `json:"number"`
+	Status     string  `json:"status"`
+	Accrual    float64 `json:"accrual"`
+	UploadedAt string  `json:"uploaded_at"`
 }
 
 type Balance struct {
